Create complete backend with essential packages
Set up the backend with Express, Mongoose, and all the key packages like bcryptjs, jsonwebtoken, cors, and dotenv.
