+++ backend/eslint.config.mjs
@@ -1,0 +1,10 @@
+import globals from "globals";
+import pluginReact from "eslint-plugin-react";
+
+
+/** @type {import('eslint').Linter.Config[]} */
+export default [
+  {files: ["**/*.{js,mjs,cjs,jsx}"]},
+  {languageOptions: { globals: {...globals.browser, ...globals.node} }},
+  pluginReact.configs.flat.recommended,
+];
