+++ frontend/src/NavBar/Login.css
@@ -1,0 +1,5 @@
+/* frontend\src\NavBar\Login.css */
+.main-container-login {
+  height: 100vh;
+  background-color: white;
+}
