+++ frontend/src/index.css
@@ -1,0 +1,11 @@
+/* frontend/src/index.css */
+body {
+  background-color: #00c48c;
+  font-family: sans-serif;
+  height: 80%;
+}
+
+.btn- {
+  width: 15rem;
+  height: 4rem;
+}
