--- src/Body/CreateList.css
@@ -1,145 +1,0 @@
-.main-c-container {
-    border: 2px solid #00C48C;
-    border-radius: 1rem;
-    padding: 1rem;
-   
-  }
-  
-  .opt-container {
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-  }
-  
-  select {
-    padding: 0.5rem;
-    border-radius: 0.5rem;
-    border: 1px solid #ccc;
-    font-size: 1rem;
-    background-color: white;
-    cursor: pointer;
-  }
-  
-  select:focus {
-    outline: none;
-    border-color: #007bff;
-    box-shadow: 0 0 5px rgba(0, 123, 255, 0.5);
-  }
-  
-  /* Optional: Style for the date div */
-  .opt-container div {
-    font-size: 1rem;
-    color: #333;
-  }
-  
-  /* Minimalist amount input */
-  .amount-text {
-    width: 10rem;
-    height: 2rem;
-    border: none;
-    border-bottom: 3px solid #00C48C; /* Subtle bottom border */
-    background-color: transparent;
-    font-size: 1.5rem;
-    color: #333; /* Dark text for contrast */
-    padding: 0.25rem 0; /* Add some vertical padding */
-    transition: border-color 0.3s ease; /* Smooth transition for focus */
-    color: orange; /* Text color */
-    text-shadow: 0 0 2px orange;
-    padding-left: 40px;
-  }
-
-  .amount-text:focus{
-    outline: none;
-  }
-  
-  
-  /* Hide number input spinners */
-  .amount-text[type="number"] {
-    -moz-appearance: textfield; /* Firefox */
-  }
-  
-  .amount{
-    background-color: #00C48C;
-    color: white;
-    padding: 10px;
-    border-radius: 1rem;
-    margin-right: 2rem;
-  }
-
-  .amount-text-container{
-   display: flex;
-  
-  }
-
-  .note-container{
-    width: 80%;
-    margin: 0 auto;
-  }
-
-  .note-text {
-    width: 30rem; /* Fixed width */
-    height: 5rem; /* Fixed height */
-    resize: none; /* Prevent resizing */
-    padding: 0.5rem; /* Add padding */
-    border: 1px solid #ccc; /* Add a border */
-    border-radius: 0.5rem; /* Rounded corners */
-    font-size: 1rem; /* Font size */
-    text-align: left; /* Align text to the left */
-    vertical-align: top; /* Start text from the top */
-  }
-  
-  .note-text:focus {
-    outline: none; /* Remove default focus outline */
-    border-color: #00C48C; /* Change border color on focus */
-    box-shadow: 0 0 5px rgba(0, 196, 140, 0.5); /* Add a subtle shadow on focus */
-  }
-
-  
-  
-  label {
-    font-size: 1rem;
-    color: #333;
-  }
-  
-  input[type="date"] {
-    padding: 0.5rem;
-    border: 1px solid #ccc;
-    border-radius: 0.5rem;
-    font-size: 1rem;
-  }
-  
-  input[type="submit"] {
-    padding: 0.5rem;
-    border: none;
-    border-radius: 0.5rem;
-    background-color: #00C48C;
-    color: white;
-    font-size: 1rem;
-    cursor: pointer;
-  }
-  
-  input[type="submit"]:hover {
-    background-color: #009973;
-  }
-
-  .createlist-add-btn {
-    background-color: #00C48C;
-    color: white;
-    padding: 0.75rem 1.5rem;
-    border: none;
-    border-radius: 0.5rem;
-    font-size: 1rem;
-    font-weight: 500;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-    display: block; /* Make the button a block-level element */
-    margin: 0 auto; /* Center horizontally */
-    width: fit-content; /* Adjust width to fit content */
-    margin-top: 20px;
-  }
-  
-  .createlist-add-btn:hover {
-    background-color: #009973; /* Darker green on hover */
-  }
-  
-
