--- src/Body/DisplayAmount.css
@@ -1,39 +1,0 @@
-.display-amount-text{
-    width: 10rem;
-    height: 3.1rem;
-    border: none;
-    border-bottom: 3px solid #00C48C; /* Subtle bottom border */
-   
-    font-size: 1.5rem;
-    color: #333; /* Dark text for contrast */
-    padding: 0.25rem 0; /* Add some vertical padding */
-    transition: border-color 0.3s ease; /* Smooth transition for focus */
-    color: orange; /* Text color */
-    text-shadow: 0 0 2px orange;
-    padding-left: 40px;
-    border-radius: 10px;
-}
-
-.display-amount-text:focus{
-    outline: none;
-  }
-
-  .display-amount-text[type="number"] {
-    -moz-appearance: textfield; /* Firefox */
-  }
-
-.amount-add-btn{
-    
-    color:  #00C48C;
-    padding: 0.75rem 1.5rem;
-    border: none;
-    border-radius: 0.5rem;
-    font-size: 1rem;
-    font-weight: 500;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-    display: block; /* Make the button a block-level element */
-    width: fit-content; /* Adjust width to fit content */
-    margin-left: 10px;
-}
-
