--- src/Body/List.css
@@ -1,77 +1,0 @@
-.date{
-    font-family: sans-serif;
-    color: #0A0A3A;
-    margin-left: 7rem;
-}
-
-.list--container{
-    border: 1px solid white;
-    background-color: #f9f9f9;
-    height: 20rem; /* Fixed height */
-    overflow-y: auto; /* Enable vertical scrolling */
-    padding: 10px; /* Optional: Add some padding */
-}
-
-.item--container{
-    display: flex;
-    justify-content: space-between;
-    padding: 0px 100px;
-
-
-}
-
-.hr-line{
-   
-    width: 80%; /* Keep the width */
-    height: 2px; /* Set a small height for the line */
-    border: 0; /* Remove the default border */
-    background-color:  #00C48C;
-    margin: 10px auto; /* Center the line and add some margin */
-      
-}
-
-.item-cost{
-    color: orange; /* Text color */
-    text-shadow: 0 0 2px orange;
-}
-
-.items{
-    margin-left: 1rem;
-}
-
-.item-note{
-    width: 80%;
-    font-size: 20px;
-    margin: 0 auto;
-    margin-bottom: 20px;
-    margin-left: 100px;
-   
-}
-
-.info-text {
-    display: flex;
-    justify-content: center; /* Centers horizontally */
-    align-items: center; /* Centers vertically */
-    height: 30vh; /* Full viewport height to center vertically in the viewport */
-    margin: 0 auto;
-}
-
-.delete-btn-list{
-    background-color: #00C48C;
-    color: white;
-    padding: 0.75rem 1.5rem;
-    border: none;
-    border-radius: 0.5rem;
-    font-size: 1rem;
-    font-weight: 500;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-    display: block; /* Make the button a block-level element */
-    margin: 0 auto; /* Center horizontally */
-    width: fit-content; /* Adjust width to fit content */
-    margin-top: 20px;
-}
-
-.delete-btn-list:hover {
-    background-color: #009973; /* Darker green on hover */
-  }
