--- src/Body/MonthList.css
@@ -1,13 +1,0 @@
-.height{
-    height: 500px;
-}
-
-.display-date{
-   text-align: center;
-   margin-top: 40px;
-}
-
-.display-date-details{
-    margin-left: 100px;
-    color: red;
-}
