--- src/Footer.css
@@ -1,20 +1,0 @@
-.footer {
-    height: 100px;
-   background-color: #f9f9f9;
-    display: flex;
-    box-shadow: 0 5px 20px rgba(0, 0, 0, 0.15);
-
-    align-items: center;
-    
-   
-}
-
-img {
-    width: 50px;
-    border-radius: 40px;
-    margin-left: 3rem;
-
-}
-
-
-
