--- src/NavBar.css
@@ -1,72 +1,0 @@
-.main-container{
-    position: sticky; /* Makes the container sticky */
-    top: 0; /* Stick to the top of the viewport */
-    z-index: 1000; /* Ensures it stays above other content */
-    width: 100%; /* Ensure it spans the full width */
-    padding: 10px 0; /* Add padding for better spacing */
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-    background-color: #00C48C;
-    
-}
-
-.container-acc{
-    display: flex;
-    justify-content: space-between;
-    margin: 0px 70px;
-}
-
-.btn{
-   margin: 0.5rem;
-   width: 5.5rem;
-   height: 3rem;
-   font-size: 1rem;
-   border-radius: 10px;
-   border: 1px solid white;
-   color:#04AA6D; 
-   margin-top: 3rem;
-}
-
-
-.btn-sign-up:hover{
-    background-color:lightblue;
-    color:#04AA6D;
-}
-
-.btn-login:hover{
-    background-color:lightblue;
-    color: #04AA6D;
-}
-
-.budget-container{
-    color: white;
-    font-family: sans-serif;
-    margin-left: 10rem;
-    margin-top: 7rem;
-}
-
-h1{
-    font-size: 3rem;
-}
-
-.welcome-text{
-    color: white;
-    margin-top: 4rem;
-}
-
-.btn-log-out{
-    margin-left: 2rem;
-    width: 5.5rem;
-    height: 3rem;
-    font-size: 1rem;
-    border-radius: 10px;
-    border: 1px solid white;
-    color:#04AA6D;
-    margin-top: 3rem;
-    
-}
-
-.btn-log-out:hover{
-    background-color:lightblue;
-    color: #04AA6D;
-   
-}
