--- src/NavBar/Login.css
@@ -1,4 +1,0 @@
-.main-container-login{
-    height: 100vh;
-    background-color: white;
-}
