--- src/NavBar/SignUp.css
@@ -1,81 +1,0 @@
-.Main-Box{   
-    background-color: #f8f9fc; /* Light background */
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.btn-back-container{
-    position: absolute;
-    top: 1rem;
-    right: 1rem;
-    display: flex;
-    justify-content: flex-end;
-}
-
-.btn-back{
-    font-size: 2rem;
-   
-}
-
-.container {
-    background-color: #fff; /* White card */
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    width: 320px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.container div:first-child {
-    font-size: 24px;
-    font-weight: 600;
-    margin-bottom: 20px;
-    color: #333;
-}
-
-input[type="text"],
-input[type="password"] {
-    padding: 10px;
-    margin-bottom: 15px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    font-size: 14px;
-    color: #333;
-}
-
-input[type="text"]:focus,
-input[type="password"]:focus {
-    outline: none;
-    border-color: #007bff;
-    box-shadow: 0 0 3px rgba(0, 123, 255, 0.5);
-}
-
-.eye {
-    position: absolute;
-    margin: 10px;
-}
-
-.btn-create {
-    width: 65%;
-    padding: 10px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    font-size: 16px;
-    cursor: pointer;
-    font-weight: 500;
-}
-
-.btn-create:hover {
-    background-color: #0056b3;
-}
-
-.btn-create:focus {
-    outline: none;
-    box-shadow: 0 0 3px rgba(0, 123, 255, 0.5);
-}
