--- src/index.css
@@ -1,13 +1,0 @@
-
-body{
-    background-color: #00C48C;
-    font-family: sans-serif;
-    height: 80%;
-    
-}
-
-.btn-{
-    width: 15rem;
-    height: 4rem;
-    
-}
